Fixing of extension, syntax and json File
This allows the extension to send messages to the language Server and
therefore the content assist now works the same like in Eclipse. Also
the validation of the pragma and the validation of the syntax works now
on the fly and not only after saving the file.
Special thanks to cdietrich