Function Signature formatting alligned to styleguide
Function signatures still not getting formatted exactly according to the
style guide but much more than before. In detail: If the function
signature is longer than 80 characters, the parameters of the function
are now listed line by line instead of multiple parameters in one line.
If the function signature is shorter than 80 characters the signature
will be formatted in one line.
Also the SemanticHighlighter had a bug which was fixed.